V2.0.0 Beta3

+++ django_wms/__init__.py
@@ -7,11 +7,15 @@ from django.conf import settings
 mimetypes.add_type("text/css", ".css", True)
 mimetypes.add_type("text/javascript", ".js", True)
 try:
-    response = requests.post('https://www.56yhz.com/area/')
+
     path = os.path.join(settings.BASE_DIR, 'utils/country.txt')
-    with open(path, 'w') as f:
-        f.write(str(eval(response.text).get('country')))
-        f.close()
+    if os.path.exists(path):
+        pass
+    else:
+        response = requests.post('https://www.56yhz.com/area/')
+        with open(path, 'w') as f:
+            f.write(str(eval(response.text).get('country')))
+            f.close()
 except:
     pass
 print('Welcome To GreaterWMS')
